+++ Projects/FeatureMemo/Targets/FeatureMemo/List/MemoListView.swift
@@ -11,6 +11,25 @@ import Core
 
 import ComposableArchitecture
 
+
+struct FileItemView: View {
+    let file: File
+    
+    init(file: File) {
+        self.file = file
+    }
+    
+    var body: some View {
+        VStack(alignment: .leading) {
+            Text(file.title)
+                .font(.headline)
+            Text(file.content)
+                .foregroundColor(.gray)
+                .font(.callout)
+        }
+    }
+}
+
 struct MemoListView: View {
     let store: StoreOf<MemoList>
     @State private var searchText = ""
